% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas-viewports.R
\name{appendPath}
\alias{appendPath}
\title{appends a vpPath to include the number n on the bottommost viewport. If the
bottom most viewport ends in a number, it replaces that number with n.}
\usage{
appendPath(vp, n)
}
\description{
appends a vpPath to include the number n on the bottommost viewport. If the
bottom most viewport ends in a number, it replaces that number with n.
}
