% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CI-mean.R
\name{calcCITWald}
\alias{calcCITWald}
\title{the various confidence coverage methods for CALC_STAT
note that the percentile bootstrap methods do not perform very well. Here's the speed of makeStatDistribution when using different version of them on an original data of size 80}
\usage{
calcCITWald(x, y = NULL)
}
\description{
the various confidence coverage methods for CALC_STAT
note that the percentile bootstrap methods do not perform very well. Here's the speed of makeStatDistribution when using different version of them on an original data of size 80
}
