% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CI-mean.R
\name{ci1000}
\alias{ci1000}
\title{confidence coverage method for HANDLE_1000: how to display the results of 1000 bootstrap samples}
\usage{
ci1000(canvas, e, fun = mean)
}
\description{
confidence coverage method for HANDLE_1000: how to display the results of 1000 bootstrap samples
}
