% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CI-mean.R
\name{CIcounter}
\alias{CIcounter}
\title{confidence coverage method for DISPLAY_RESULT}
\usage{
CIcounter(canvas, env, cov.message = TRUE, fun = mean)
}
\description{
confidence coverage method for DISPLAY_RESULT
}
