% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-coloureddatatext.R
\name{coldatatextGrob}
\alias{coldatatextGrob}
\title{Construct a datatext grob}
\usage{
coldatatextGrob(data, title = NULL, cols = NULL, xpos = 0.5, max = 50,
  name = NULL, gp = NULL, vp = NULL, dp = 2)
}
\arguments{
\item{data}{a numeric or character vector of data.}

\item{title}{a character string to provide the title, usually the name of the data vector.}

\item{cols}{a vector of character strings specifying the colours for each level of a factor.}

\item{name}{a name for the grob to be constructed.}

\item{gp}{graphical parameters for the text, constructed with gpar().}

\item{vp}{a default viewport to be used when drawing the grob.}
}
\description{
datatextGrob constructs text of a vector of data with colours for different levels of a factor.
datatextGrobs inherit the class "coldatatext".
}

