% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-ci.R
\name{confintGrob}
\alias{confintGrob}
\title{Construct a ci grob}
\usage{
confintGrob(
  ci,
  at = unit(0.15, "npc"),
  height = unit(0.01, "npc"),
  to = unit(-2, "lines"),
  col = "red",
  bar = TRUE,
  cilabs = ci,
  fix.label.pos = c(FALSE, FALSE),
  name = NULL,
  gp = NULL,
  vp = NULL
)
}
\arguments{
\item{ci}{the confidence interval to be plotted, given by a vector of length 2.}

\item{at}{a unit object giving the y-axis position to plot the confidence interval.}

\item{to}{a unit object giving the y-axis positions of where the arrows are to point.}

\item{col}{the colour to plot the confidence interval.}

\item{bar}{a logical value determining whether or not a bar is plotted.}
}
\description{
datatextGrob constructs text of a vector of data.
datatextGrobs inherit the class "confint".
}
