% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-datatext.R
\name{datatextGrob}
\alias{datatextGrob}
\title{Construct a datatext grob}
\usage{
datatextGrob(
  data,
  title = NULL,
  x = 0.5,
  max = 50,
  name = NULL,
  gp = NULL,
  vp = NULL,
  dp = 2
)
}
\arguments{
\item{data}{a numeric or character vector of data.}

\item{title}{a character string to provide the title, usually the name of the data vector.}

\item{name}{a name for the grob to be constructed.}

\item{gp}{graphical parameters for the text, constructed with gpar().}

\item{vp}{a default viewport to be used when drawing the grob.}
}
\description{
datatextGrob constructs text of a vector of data.
datatextGrobs inherit the class "datatext".
}
