% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CI-mean.R
\name{dropPoints1d}
\alias{dropPoints1d}
\title{Animates a sample of points dropping down from the collection of points in the data window. The ANIMATE_SAMPLE method for numeric, one dimensional data.}
\usage{
dropPoints1d(canvas, n.steps, n.slow, keep.plot, move = TRUE)
}
\description{
Animates a sample of points dropping down from the collection of points in the data window. The ANIMATE_SAMPLE method for numeric, one dimensional data.
}
