% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-methods-1d-numeric.R
\name{plotBoxplot}
\alias{plotBoxplot}
\title{helper function for plotting numeric data}
\usage{
plotBoxplot(canvas, x, stat, stat.color, vp, name, show.w = TRUE)
}
\description{
helper function for plotting numeric data
}

