% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-methods-1d-numeric.R
\name{plotHist}
\alias{plotHist}
\title{helper function for plotting numeric data}
\usage{
plotHist(canvas, x, vp, name, fill = "grey")
}
\description{
helper function for plotting numeric data
}
