% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-methods-1d-numeric.R
\name{plotPoints}
\alias{plotPoints}
\title{helper function for plotting numeric data}
\usage{
plotPoints(canvas, x, y, vp, name, black = FALSE, alpha = 1,
  col = "grey60")
}
\description{
helper function for plotting numeric data
}

