% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-methods-1d-numeric.R
\name{plotPointsAndBoxplot}
\alias{plotPointsAndBoxplot}
\title{Used for displaying numeric data. Plots a group of points with a boxplot in
the specified viewport. The points will be stacked on top of each other in an
intelligent manner when necessary.}
\usage{
plotPointsAndBoxplot(canvas)
}
\arguments{
\item{canvas}{a canvasClass object made with the generator object canvas
PLOT_DATA method for numeric, 1d data. Also used for 2d data when one varible is numeric}
}
\description{
Used for displaying numeric data. Plots a group of points with a boxplot in
the specified viewport. The points will be stacked on top of each other in an
intelligent manner when necessary.
}
