% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackPoints.R
\name{stackPoints}
\alias{stackPoints}
\title{Calculates y values for a vector of overlapping x values. Y values are designed to prevent overlapping by creating the appearance of stacked data. A helper function for numeric 1d Data. Also used for 2d data where one dimension is numeric, and one is categorical.}
\usage{
stackPoints(
  x,
  levels = rep(1, length(x)),
  vp,
  y.min = 0.3,
  y.max = unit(1, "npc")
)
}
\description{
Calculates y values for a vector of overlapping x values. Y values are designed to prevent overlapping by creating the appearance of stacked data. A helper function for numeric 1d Data. Also used for 2d data where one dimension is numeric, and one is categorical.
}
